requirements.txt
apache-airflow-providers-amazon==9.15.0
boto3==1.40.45
requests==2.32.5
python-dotenv==1.1.1
pandas==2.3.3
black==24.4.2
isort==5.13.2
pytest==8.2.0
